Commit initial dependency work.
This commit allows users to do the following:

```
/snyk dependencies list <Project Name>
```

As soon as there was a success doing the above I hit the commit button,
so this commit is quite messy with lots of excess log statements and
superfluous code from different ideas.

This commit also adds an alternative callSnykAPI() function that does
basically the same thing as the original, but takes a single argument
object instead because I hate remembering arg order.